# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Agent Framework API",
        "title": "GoAgent API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
        "/agentos/plans": {
            "post": {
                "description": "Start a durable cross-backend AgentOS RunPlan.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "agentos"
                ],
                "summary": "Start AgentOS plan",
                "operationId": "agentos-start-plan",
                "parameters": [
                    {
                        "description": "AgentOS run plan request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_TekkenSteve_GoAgent_agentos_control.RunPlanSpec"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/github_com_TekkenSteve_GoAgent_agentos_control.RunPlanStatus"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            }
        },
...
        "/agentos/plans/{plan_id}/artifacts": {
            "get": {
                "description": "Query durable artifact refs for a RunPlan.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "agentos"
                ],
                "summary": "List AgentOS plan artifacts",
                "operationId": "agentos-list-plan-artifacts",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Plan ID",
                        "name": "plan_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Account ID",
                        "name": "account_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Project ID",
                        "name": "project_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Node ID",
                        "name": "node_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Child run ID",
                        "name": "run_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Maximum refs",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_TekkenSteve_GoAgent_agentos_core.ArtifactRef"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            }
        },
        "/agentos/plans/{plan_id}/artifacts/{artifact_id}": {
            "get": {
                "description": "Read one durable artifact document for a RunPlan.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "agentos"
                ],
                "summary": "Get AgentOS plan artifact",
                "operationId": "agentos-get-plan-artifact",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Plan ID",
                        "name": "plan_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Artifact ID",
                        "name": "artifact_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Account ID",
                        "name": "account_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Project ID",
                        "name": "project_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_TekkenSteve_GoAgent_agentos_core.Artifact"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            }
        },
        "/agentos/plans/{plan_id}/audits": {
            "get": {
                "description": "Query durable audit records for plan control-plane actions.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "agentos"
                ],
                "summary": "List AgentOS plan audits",
                "operationId": "agentos-list-plan-audits",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Plan ID",
                        "name": "plan_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Account ID",
                        "name": "account_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Project ID",
                        "name": "project_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Node ID",
                        "name": "node_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Child run ID",
                        "name": "run_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Audit action",
                        "name": "action",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Maximum records",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_TekkenSteve_GoAgent_agentos_control.PlanAuditRecord"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            }
        },
...
        "/agentos/plans/{plan_id}/control": {
            "post": {
                "description": "Send lifecycle control such as pause, resume, or cancel to an AgentOS RunPlan.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "agentos"
                ],
                "summary": "Control AgentOS plan",
                "operationId": "agentos-control-plan",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Plan ID",
                        "name": "plan_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "AgentOS plan control operation",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.AgentOSPlanControl"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            }
        },
...
        "/agentos/plans/{plan_id}/description": {
            "get": {
                "description": "Query the current public topology and aggregate status of an AgentOS RunPlan.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "agentos"
                ],
                "summary": "Describe AgentOS plan",
                "operationId": "agentos-plan-description",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Plan ID",
                        "name": "plan_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Account ID",
                        "name": "account_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Project ID",
                        "name": "project_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_TekkenSteve_GoAgent_agentos_control.RunPlanDescription"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_TekkenSteve_GoAgent_agentos_control.RunPlanSpec": {
            "type": "object",
            "properties": {
                "account_id": {
                    "type": "string"
                },
                "edges": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/control.PlanEdgeSpec"
                    }
                },
                "idempotency_key": {
                    "type": "string"
                },
                "inputs": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "nodes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/control.PlanNodeSpec"
                    }
                },
                "plan_id": {
                    "type": "string"
                },
                "policy": {
                    "$ref": "#/definitions/control.PlanPolicy"
                },
                "project_id": {
                    "type": "string"
                },
                "requested_at": {
                    "type": "string"
                },
                "thread_id": {
                    "type": "string"
                }
            }
        },
        "github_com_TekkenSteve_GoAgent_agentos_control.RunPlanStatus": {
            "type": "object",
            "properties": {
                "active_run_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "artifacts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/core.ArtifactRef"
                    }
                },
                "budget_usage": {
                    "$ref": "#/definitions/control.PlanBudgetUsage"
                },
                "lifecycle_state": {
                    "type": "string"
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "nodes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/control.PlanNodeStatus"
                    }
                },
                "plan_id": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                },
                "started_at": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "response.Error": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "message"
                }
            }
        },
...
    }
}